// src/models/Farmer.ts
import {Schema, model} from 'mongoose'
import Farmer from '../interfaces/Farmer'

const farmerSchema = new Schema<Farmer>({
  lakshmiCentreCode: {
    type: String,
  },
  // This will be the unique code generated by the stell apps and will be unique across the whole platform
  uid: {
    type: String,
    // unique: true,
  },
  firstName: {
    type: String,
  },
  lastName: {
    type: String,
  },
  // password: {
  //   type: String,
  // },
  profileImage: {
    type: String,
  },
  email: {
    type: String,
  },
  address: {
    locality: String,
    pincode: String,
    district: String,
    taluka: String,
    state: String,
    country: String,
  },
  location: {home: [Number, Number], farm: [Number, Number]},
  gender: {type: String, enum: ['male', 'female']},
  approvalStatus: {type: String, enum: ['approved', 'refused', 'pending']},
  phoneNumber: {
    type: String,
    unique: true,
    required: true,
  },
  preferredLanguage: {
    type: String,
    enum: ['marathi', 'english'],
    default: 'marathi',
  },
  dateOfBirth: {
    type: Date,
  },
  age: {
    type: Number,
  },
  /**
   * Date of registration on app
   */
  registrationDate: {
    type: Date,
  },

  /**
   * Date of starting as a Milk pouring farmer
   */
  activeSince: {
    type: Date,
  },

  /**
   * Profile is created/managed now
   */
  managedBy: {
    type: String,
    enum: ['self', 'admin', 'agent'],
    required: true,
  },
  coins: {
    type: String,
  },

  phoneOtp: [
    {
      val: Number,
      exp: Number,
    },
  ],
})

const FarmerModel = model<Farmer>('Farmer', farmerSchema)

export default FarmerModel